requirements.txt
requests
mycroft-messagebus-client>=0.8.4
ovos_utils==0.0.8a3
python-vlc==1.1.2
neon-utils @ git+https://github.com/neongeckocom/neon-skill-utils
neon_enclosure @ git+https://github.com/neongeckocom/neon-enclosure
mycroft-lib @ git+https://github.com/HelloChatterbox/mycroft-lib